grin2 route calls getter instead of reading file, moved route code
